Explicitly cast in a few places to not lose precision during arithmetic

+++ tests/avifyuv.c
@@ -215,14 +215,14 @@ int main(int argc, char * argv[])
 
                 printf(" * depth: %d, matrixCoeffs: %d, maxDrift: %2d\n", yuvDepth, matrixCoeffs, maxDrift);
 
-                const size_t totalPixelCount = dim * dim * dim;
+                const size_t totalPixelCount = (size_t)dim * dim * dim;
                 for (int i = 0; i < MAX_DRIFT; ++i) {
                     if (verbose && (driftPixelCounts[i] > 0)) {
                         printf("   * drift: %2d -> %11zu / %11zu pixels (%.2f %%)\n",
                                i,
                                driftPixelCounts[i],
                                totalPixelCount,
-                               (float)driftPixelCounts[i] * 100.0f / totalPixelCount);
+                               (double)driftPixelCounts[i] * 100.0 / (double)totalPixelCount);
                     }
                 }
